% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_filtering.R
\name{data_filtering}
\alias{data_filtering}
\title{Data Filtering}
\usage{
data_filtering(x)
}
\arguments{
\item{x}{A data matrix  object with rows
including proteins and fractions along the columns.}
}
\value{
Filtered matrix.
}
\description{
This function removes proteins for which peptide only
detected in one fraction (i.e., "one-hit-wonders") across the
co-elution table, common contaminants (e.g., keratins) only for mouse
and human organisms and frequent flyers.
}
\details{
data_filtering
}
\examples{
# Load the co-elution data
data("exampleData")
# Perform raw data pre-processing
datOut <- data_filtering(exampleData)
}
\author{
Matineh Rahmatbakhsh, \email{matinerb.94@gmail.com}
}
