% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exampleData.R
\docType{data}
\name{exampleData}
\alias{exampleData}
\title{Demo CF-MS data}
\format{
A matrix with 284 rows and 83 columns,
with proteins in rows and biochemical fractions in columns
}
\usage{
data(exampleData)
}
\description{
Co-elution profiles derived from mitochondrial (mt) extracts of
mouse brain culture,fractionated by (size-exclusion chromatography, SEC)
}
