% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCPX.R
\name{getCPX}
\alias{getCPX}
\title{Fetch Complexes from the CORUM Database}
\usage{
getCPX(org = "Mouse", tpath = tempdir())
}
\arguments{
\item{org}{Mammalian model organisms including: "Human", "Mouse",
"Pig", "Bovine", "Rat", "Mammalia", "Rabbit", "Dog", "Hamster",
and "MINK". Defaults to "Mouse".}

\item{tpath}{A character string indicating the path to the project
directory. If the directory is
missing, it will be stored in the temp directory.}
}
\value{
A list containing protein complexes for mammalian organisms.
}
\description{
This function retrieves protein complexes directly from the
CORUM database.
}
\details{
getCPX
}
\author{
Matineh Rahmatbakhsh, \email{matinerb.94@gmail.com}
}
