% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refcpx.R
\docType{data}
\name{refcpx}
\alias{refcpx}
\title{CORUM reference complexes}
\source{
\url{https://mips.helmholtz-muenchen.de/corum/}
}
\usage{
data(refcpx)
}
\description{
A list containing CORUM reference complexes for mouse
organism.
}
